+++ examples/dart_examples/bin/sqlite.dart
@@ -5,14 +5,16 @@ import 'package:sqlite_async/sqlite3.dart';
 import 'package:sqlite_async/sqlite_async.dart';
 
 final migrations = SqliteMigrations()
-  ..add(SqliteMigration(1, (tx) async {
-    await tx.execute(
-      '''CREATE TABLE test_data(
+  ..add(
+    SqliteMigration(1, (tx) async {
+      await tx.execute(
+        '''CREATE TABLE test_data(
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         data TEXT
       )''',
-    );
-  }));
+      );
+    }),
+  );
 
 typedef TestData = ({int id, String data});
 final items = listSignal<TestData>([]);
@@ -26,12 +28,16 @@ void main() async {
         'SELECT * FROM test_data',
         triggerOnTables: {'test_data'},
       )
-      .map((rows) => rows
-          .map((e) => (
+      .map(
+        (rows) => rows
+            .map(
+              (e) => (
                 id: e['id'] as int,
                 data: e['data'] as String,
-              ))
-          .toList())
+              ),
+            )
+            .toList(),
+      )
       .listen((event) => items.value = event);
 
   final dispose = effect(() {
@@ -40,7 +46,7 @@ void main() async {
 
   await db.executeBatch('INSERT INTO test_data(data) values(?)', [
     ['Test1'],
-    ['Test2']
+    ['Test2'],
   ]);
 
   await db.writeTransaction((tx) async {