+++ packages/preact_signals/lib/src/signal.dart
@@ -66,6 +66,9 @@ class Signal<T> with ReadonlySignal<T> {
   }
 
   @override
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  @pragma('wasm:prefer-inline')
   T get value {
     final node = addDependency();
     if (node != null) {
@@ -91,9 +94,12 @@ class Signal<T> with ReadonlySignal<T> {
   }
 
   @internal
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  @pragma('wasm:prefer-inline')
   void internalSetValue(T val) {
     if (batchIteration > 100) {
-      throw Exception('Cycle detected');
+      throwCycleDetected();
     }
 
     internalValue = val;