+++ packages/signals_hooks/lib/src/base.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:signals_flutter/signals_core.dart';
+import 'package:signals/signals_core.dart';
 
 class SignalHook<T, S extends ReadonlySignal<T>> extends Hook<S> {
   const SignalHook(this.target, {super.keys});
@@ -24,6 +24,7 @@ class SignalHookState<T, S extends ReadonlySignal<T>>
   @override
   void dispose() {
     _cleanup?.call();
+    super.dispose();
   }
 
   void _listener() {